Move Tag classes to a new package
