+++ src/seedu/addressbook/parser/Parser.java
@@ -2,8 +2,8 @@ package seedu.addressbook.parser;
 
 import seedu.addressbook.commands.*;
 import seedu.addressbook.model.IllegalValueException;
-import seedu.addressbook.model.Tag;
-import seedu.addressbook.model.UniqueTagList;
+import seedu.addressbook.model.tag.Tag;
+import seedu.addressbook.model.tag.UniqueTagList;
 import seedu.addressbook.model.person.*;
 
 import java.util.Arrays;
