+++ src/seedu/addressbook/storage/jaxb/AdaptedAddressBook.java
@@ -2,8 +2,8 @@ package seedu.addressbook.storage.jaxb;
 
 import seedu.addressbook.model.AddressBook;
 import seedu.addressbook.model.IllegalValueException;
-import seedu.addressbook.model.Tag;
-import seedu.addressbook.model.UniqueTagList;
+import seedu.addressbook.model.tag.Tag;
+import seedu.addressbook.model.tag.UniqueTagList;
 import seedu.addressbook.model.person.Person;
 import seedu.addressbook.model.person.ReadOnlyPerson;
 import seedu.addressbook.model.person.UniquePersonList;
