+++ src/seedu/addressbook/storage/jaxb/AdaptedPerson.java
@@ -2,8 +2,8 @@ package seedu.addressbook.storage.jaxb;
 
 import seedu.addressbook.common.Utils;
 import seedu.addressbook.model.IllegalValueException;
-import seedu.addressbook.model.Tag;
-import seedu.addressbook.model.UniqueTagList;
+import seedu.addressbook.model.tag.Tag;
+import seedu.addressbook.model.tag.UniqueTagList;
 import seedu.addressbook.model.person.*;
 
 import javax.xml.bind.annotation.XmlAttribute;
