+++ src/seedu/addressbook/storage/jaxb/AdaptedTag.java
@@ -2,7 +2,7 @@ package seedu.addressbook.storage.jaxb;
 
 import seedu.addressbook.common.Utils;
 import seedu.addressbook.model.IllegalValueException;
-import seedu.addressbook.model.Tag;
+import seedu.addressbook.model.tag.Tag;
 
 import javax.xml.bind.annotation.XmlValue;
