+++ test/java/seedu/addressbook/parser/ParserTest.java
@@ -4,8 +4,8 @@ import org.junit.Before;
 import org.junit.Test;
 import seedu.addressbook.commands.*;
 import seedu.addressbook.model.IllegalValueException;
-import seedu.addressbook.model.Tag;
-import seedu.addressbook.model.UniqueTagList;
+import seedu.addressbook.model.tag.Tag;
+import seedu.addressbook.model.tag.UniqueTagList;
 import seedu.addressbook.model.person.*;
 
 import java.util.Arrays;
